Swap package and definition name
It makes more sense to have the schema /version/ (discovered from a
target .cue filename prefix) as the CUE package name, and the final
directory component (nominally the "thing" we're formalising as a
schema) encoded as the CUE definition name.

Before this commit, catalog entries would be used as follows:

    import "cue.schemastore.org/v0/some_org/some_project/some_schema"

    something: some_schema.#V1_2_3 & {
      ...
    }

This has 2 main disadvantages:

- A schema version upgrade must be orchestrated by making changes
  throughout each .cue file, at each point of use;
- Every version is brought into scope at the moment of import, even
  though the 99% use-case will be the use of a single version.

By making this change, and going forward, catalog entries will be used
like this:

    import s "cue.schemastore.org/v0/some_org/some_project/some_schema:v1_2_3"

    something: s.#SomeSchema & {
      ...
    }

Note the `s` import alias which (whilst it's only /required/ if a single
.cue file imports multiple schemas which happen to share the same
version number) produces a nicer DevX when it's present. This approach
also allows for version upgrades to be orchestrated more simply, at the
import statement level - which might even be made into a mutable element
by CUE tooling in the future. Lastly, only 1 schema version is now in
scope to be evaluated, which should be more efficient as the versioned
corpus increases.